Incorporate sampling visualisation script into CLI refactor
Renamed this to "euphonic-show-sampling", added basic docs
