+++ doc/source/tutorials.rst
@@ -32,6 +32,7 @@ Helper Scripts
 - :ref:`dispersion.py <disp-script>`
 - :ref:`dos.py <dos-script>`
 - :ref:`optimise_eta.py <eta-script>`
+- :ref:`euphonic-show-sampling <sampling-script>`
 
 Examples
 ========
